+++ resources/org/javarosa/core/util/area_with_less_than_three_points.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0"?>
+<h:html xmlns="http://www.w3.org/2002/xforms" xmlns:ev="http://www.w3.org/2001/xml-events" xmlns:h="http://www.w3.org/1999/xhtml" xmlns:jr="http://openrosa.org/javarosa" xmlns:orx="http://openrosa.org/xforms/" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+    <h:head>
+        <h:title>Area with less than three points</h:title>
+        <model>
+            <instance>
+                <area id="area" version="2018042401">
+                    <geoshape1>38.253094215699576 21.756382658677467;38.25021274773806 21.756382658677467</geoshape1>
+                    <arearesult/>
+                    <distanceresult/>
+                    <meta>
+                        <instanceID/>
+                    </meta>
+                </area>
+            </instance>
+            <bind nodeset="/area/geoshape1" type="geoshape"/>
+            <bind calculate="enclosed-area( /area/geoshape1 )" nodeset="/area/arearesult" type="string"/>
+            <bind calculate="concat('uuid:', uuid())" nodeset="/area/meta/instanceID" readonly="true()" type="string"/>
+        </model>
+    </h:head>
+    <h:body>
+        <input ref="/area/geoshape1">
+            <label>Draw your shape here...</label>
+        </input>
+    </h:body>
+</h:html>
