+++ resources/org/javarosa/core/util/distance_with_less_than_two_points.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0"?>
+<h:html xmlns="http://www.w3.org/2002/xforms" xmlns:ev="http://www.w3.org/2001/xml-events"
+        xmlns:h="http://www.w3.org/1999/xhtml" xmlns:jr="http://openrosa.org/javarosa"
+        xmlns:orx="http://openrosa.org/xforms/" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+    <h:head>
+        <h:title>Distance with less than two points</h:title>
+        <model>
+            <instance>
+                <distance id="distance" version="2018042401">
+                    <geotrace>0 0</geotrace>
+                    <trace-result/>
+                    <meta>
+                        <instanceID/>
+                    </meta>
+                </distance>
+            </instance>
+            <bind nodeset="/distance/geotrace" type="geotrace"/>
+            <bind calculate="distance( /distance/geotrace     )" nodeset="/distance/trace-result" type="string"/>
+            <bind calculate="concat('uuid:', uuid())" nodeset="/distance/meta/instanceID" readonly="true()" type="string"/>
+        </model>
+    </h:head>
+    <h:body>
+        <repeat nodeset="/distance/geotrace"/>
+    </h:body>
+</h:html>
